requirements.txt
streamlit>=1.36.0
opencv-python>=4.9.0.80
numpy>=1.26.0
Pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.17.0